Move unit tests inside unit/ folder
