+++ glean/package.json
@@ -35,12 +35,12 @@
     "glean": "./dist/cli/cli.js"
   },
   "scripts": {
-    "test": "npm run test:core && npm run test:platform && npm run test:plugins",
-    "test:core": "npm run test:base -- \"tests/core/**/*.spec.ts\" --recursive",
-    "test:plugins": "npm run test:base -- \"tests/plugins/**/*.spec.ts\" --recursive",
-    "test:platform": "npm run build:test-webext && npm run test:base -- \"tests/platform/**/*.spec.ts\" --recursive --timeout 0",
+    "test:unit": "npm run test:unit:core && npm run test:unit:platform && npm run test:unit:plugins",
+    "test:unit:core": "npm run test:base -- \"tests/unit/core/**/*.spec.ts\" --recursive",
+    "test:unit:plugins": "npm run test:base -- \"tests/unit/plugins/**/*.spec.ts\" --recursive",
+    "test:unit:platform": "npm run build:test-webext && npm run test:base -- \"tests/unit/platform/**/*.spec.ts\" --recursive --timeout 0",
     "test:base": "node --experimental-modules --experimental-specifier-resolution=node --loader=ts-node/esm node_modules/mocha/lib/cli/cli.js",
-    "build:test-webext": "cd tests/platform/utils/webext/sample/ && npm install && npm run build:xpi",
+    "build:test-webext": "cd tests/unit/platform/utils/webext/sample/ && npm install && npm run build:xpi",
     "lint": "eslint . --ext .ts,.js,.json --max-warnings=0",
     "lint:circular-deps": "madge --circular src/ --extensions ts",
     "fix": "eslint . --ext .ts,.js,.json --fix",
@@ -86,6 +86,7 @@
     "gh-pages": "^3.1.0",
     "jsdom": "16.5.3",
     "jsdom-global": "3.0.2",
+    "jsonschema": "^1.4.0",
     "madge": "^4.0.2",
     "mocha": "^8.3.0",
     "selenium-webdriver": "^4.0.0-alpha.8",
