+++ glean/tests/unit/core/glean.spec.ts
@@ -5,18 +5,18 @@
 import assert from "assert";
 import sinon from "sinon";
 
-import { CLIENT_INFO_STORAGE, DELETION_REQUEST_PING_NAME, KNOWN_CLIENT_ID } from "../../src/core/constants";
-import CoreEvents from "../../src/core/events";
-import Glean from "../../src/core/glean";
-import StringMetricType from "../../src/core/metrics/types/string";
-import CounterMetricType from "../../src/core/metrics/types/counter";
-import PingType from "../../src/core/pings/ping_type";
-import type { JSONObject } from "../../src/core/utils";
-import { isObject } from "../../src/core/utils";
-import TestPlatform from "../../src/platform/test";
-import Plugin from "../../src/plugins";
-import { Lifetime } from "../../src/core/metrics/lifetime";
-import { Context } from "../../src/core/context";
+import { CLIENT_INFO_STORAGE, DELETION_REQUEST_PING_NAME, KNOWN_CLIENT_ID } from "../../../src/core/constants";
+import CoreEvents from "../../../src/core/events";
+import Glean from "../../../src/core/glean";
+import StringMetricType from "../../../src/core/metrics/types/string";
+import CounterMetricType from "../../../src/core/metrics/types/counter";
+import PingType from "../../../src/core/pings/ping_type";
+import type { JSONObject } from "../../../src/core/utils";
+import { isObject } from "../../../src/core/utils";
+import TestPlatform from "../../../src/platform/test";
+import Plugin from "../../../src/plugins";
+import { Lifetime } from "../../../src/core/metrics/lifetime";
+import { Context } from "../../../src/core/context";
 
 class MockPlugin extends Plugin<typeof CoreEvents["afterPingCollection"]> {
   constructor() {
