+++ glean/tests/unit/core/metrics/datetime.spec.ts
@@ -5,12 +5,12 @@
 import assert from "assert";
 import sinon from "sinon";
 
-import Glean from "../../../src/core/glean";
-import DatetimeMetricType, { DatetimeMetric } from "../../../src/core/metrics/types/datetime";
+import Glean from "../../../../src/core/glean";
+import DatetimeMetricType, { DatetimeMetric } from "../../../../src/core/metrics/types/datetime";
 
-import TimeUnit from "../../../src/core/metrics/time_unit";
-import { Lifetime } from "../../../src/core/metrics/lifetime";
-import { Context } from "../../../src/core/context";
+import TimeUnit from "../../../../src/core/metrics/time_unit";
+import { Lifetime } from "../../../../src/core/metrics/lifetime";
+import { Context } from "../../../../src/core/context";
 
 const sandbox = sinon.createSandbox();
