+++ glean/tests/unit/core/metrics/events_database.spec.ts
@@ -3,12 +3,12 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import assert from "assert";
-import Glean from "../../../src/core/glean";
+import Glean from "../../../../src/core/glean";
 
-import { Lifetime } from "../../../src/core/metrics/lifetime";
-import EventsDatabase, { RecordedEvent } from "../../../src/core/metrics/events_database";
-import EventMetricType from "../../../src/core/metrics/types/event";
-import type { JSONObject } from "../../../src/core/utils";
+import { Lifetime } from "../../../../src/core/metrics/lifetime";
+import EventsDatabase, { RecordedEvent } from "../../../../src/core/metrics/events_database";
+import EventMetricType from "../../../../src/core/metrics/types/event";
+import type { JSONObject } from "../../../../src/core/utils";
 
 describe("EventsDatabase", function() {
   const testAppId = `gleanjs.test.${this.title}`;
