+++ glean/tests/unit/core/metrics/uuid.spec.ts
@@ -5,11 +5,11 @@
 import assert from "assert";
 import { v4 as UUIDv4 } from "uuid";
 
-import Glean from "../../../src/core/glean";
-import UUIDMetricType from "../../../src/core/metrics/types/uuid";
-import { Lifetime } from "../../../src/core/metrics/lifetime";
-import { Context } from "../../../src/core/context";
-import { ErrorType } from "../../../src/core/error/error_type";
+import Glean from "../../../../src/core/glean";
+import UUIDMetricType from "../../../../src/core/metrics/types/uuid";
+import { Lifetime } from "../../../../src/core/metrics/lifetime";
+import { Context } from "../../../../src/core/context";
+import { ErrorType } from "../../../../src/core/error/error_type";
 
 describe("UUIDMetric", function() {
   const testAppId = `gleanjs.test.${this.title}`;
