+++ glean/tests/unit/core/upload/index.spec.ts
@@ -6,13 +6,13 @@ import assert from "assert";
 import sinon from "sinon";
 import { v4 as UUIDv4 } from "uuid";
 
-import { Configuration } from "../../../src/core/config";
-import { Context } from "../../../src/core/context";
-import Glean from "../../../src/core/glean";
-import PingType from "../../../src/core/pings/ping_type";
-import { collectAndStorePing } from "../../../src/core/pings/maker";
-import PingUploader from "../../../src/core/upload";
-import { UploadResultStatus } from "../../../src/core/upload/uploader";
+import { Configuration } from "../../../../src/core/config";
+import { Context } from "../../../../src/core/context";
+import Glean from "../../../../src/core/glean";
+import PingType from "../../../../src/core/pings/ping_type";
+import { collectAndStorePing } from "../../../../src/core/pings/maker";
+import PingUploader from "../../../../src/core/upload";
+import { UploadResultStatus } from "../../../../src/core/upload/uploader";
 
 const sandbox = sinon.createSandbox();
