+++ glean/tests/unit/platform/utils/webext/sample/README.md
