+++ glean/tests/unit/platform/utils/webext/sample/background.js
