+++ glean/tests/unit/platform/utils/webext/sample/content.js
