+++ glean/tests/unit/platform/utils/webext/sample/manifest.json
