+++ glean/tests/unit/platform/utils/webext/sample/package.json
