+++ glean/tests/unit/platform/utils/webext/sample/webpack.config.js
