+++ glean/tests/unit/platform/utils/webext/sample/zip.lst
