+++ glean/tests/unit/plugins/encryption.spec.ts
@@ -8,18 +8,18 @@ import generateKeyPair from "jose/util/generate_key_pair";
 import fromKeyLike from "jose/jwk/from_key_like";
 import compactDecrypt from "jose/jwe/compact/decrypt";
 
-import Glean from "../../src/core/glean";
-import PingType from "../../src/core/pings/ping_type";
-import type { JSONObject } from "../../src/core/utils";
-import TestPlatform from "../../src/platform/test";
-import PingEncryptionPlugin from "../../src/plugins/encryption";
-import collectAndStorePing, { makePath } from "../../src/core/pings/maker";
-import type { UploadResult} from "../../src/core/upload/uploader";
-import type Uploader from "../../src/core/upload/uploader";
-import { UploadResultStatus } from "../../src/core/upload/uploader";
-import CounterMetricType from "../../src/core/metrics/types/counter";
-import { Lifetime } from "../../src/core/metrics/lifetime";
-import { Context } from "../../src/core/context";
+import Glean from "../../../src/core/glean";
+import PingType from "../../../src/core/pings/ping_type";
+import type { JSONObject } from "../../../src/core/utils";
+import TestPlatform from "../../../src/platform/test";
+import PingEncryptionPlugin from "../../../src/plugins/encryption";
+import collectAndStorePing, { makePath } from "../../../src/core/pings/maker";
+import type { UploadResult} from "../../../src/core/upload/uploader";
+import type Uploader from "../../../src/core/upload/uploader";
+import { UploadResultStatus } from "../../../src/core/upload/uploader";
+import CounterMetricType from "../../../src/core/metrics/types/counter";
+import { Lifetime } from "../../../src/core/metrics/lifetime";
+import { Context } from "../../../src/core/context";
 
 const sandbox = sinon.createSandbox();
